requirements.txt
matplotlib==3.3.4
keras==2.7.0
tensorflow-gpu==2.7.0
opencv-python==4.5.4
scikit-learn==0.24.1
seaborn==0.11.1
Pillow==8.2.0
streamlit==1.5.1
streamlit-option-menu==0.3.2